Fixing the error in console, fixing the shout name for players with no pvp title and adding the banreason

+++ src/mcMMOFix/ExpFarmFix.java
@@ -106,7 +106,7 @@ public class ExpFarmFix implements Listener
 
 		final LivingEntity target = (LivingEntity) defender;
 
-		if (CombatUtils.isInvincible(target, (int) damage))
+		if (CombatUtils.isInvincible(target, damage))
 			return false;
 
 		Entity attacker = event.getDamager();
